requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
scikit-learn
scikit-multilearn
joblib
python-multipart
